server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
PyJWT==2.8.0
SQLAlchemy==2.0.26
Werkzeug==2.3.7
python-dotenv==1.0.0
nltk==3.8.1
regex==2023.12.25
google-generativeai==0.3.1
requests==2.31.0
deep-translator==1.11.4 